requirements.txt
numpy==1.24.4
opencv-python==4.9.0.80
tqdm==4.66.2
pillow==10.3.0